requirements.txt
apipkg==1.5
appdirs==1.4.4
appnope==0.1.2
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
boto==2.49.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
commonmark==0.9.1
coverage==5.5
cryptography==3.3.2
cssselect==1.1.0
decorator==5.0.9
distlib==0.3.2
execnet==1.8.1
fasteners==0.16.1
filelock==3.0.12
idna==2.10
iniconfig==1.1.1
ipdb==0.13.9
ipython==7.24.1
ipython-genutils==0.2.0
jedi==0.18.0
jsonref==0.2
jsonschema==3.2.0
matplotlib-inline==0.1.2
more-itertools==8.8.0
msedge-selenium-tools==3.141.3
nose==1.3.7
packaging==20.9
parameterized==0.8.1
parso==0.8.2
pdfminer.six==20201018
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pluggy==0.13.1
prompt-toolkit==3.0.18
protobuf==3.17.2
ptyprocess==0.7.0
py==1.10.0
pycparser==2.20
Pygments==2.9.0
PyMySQL==1.0.2
pyOpenSSL==20.0.1
pyotp==2.6.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.4
pytest-cov==2.12.1
pytest-forked==1.3.0
pytest-html==2.0.1
pytest-metadata==1.11.0
pytest-ordering==0.6
pytest-rerunfailures==10.0
pytest-xdist==2.2.1
PyYAML==5.4.1
requests==2.25.1
rich==10.2.2
selenium==3.141.0
seleniumbase==1.63.10
setuptools-scm==6.0.1
simplejson==3.17.3
six==1.16.0
slack-sdk==3.7.0
sortedcontainers==2.4.0
soupsieve==2.2.1
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.5
virtualenv==20.4.7
wcwidth==0.2.5
